not working routing before send to mainpc
