+++ src/main.jsx
@@ -1,21 +1,21 @@
-// import React from 'react'
-// import ReactDOM from 'react-dom/client'
-// import App from './App.jsx'
-// import './index.css'
+ import React from 'react'
+ import ReactDOM from 'react-dom/client'
+ import App from './App.jsx'
+ import './index.css'
 
-// ReactDOM.createRoot(document.getElementById('root')).render(
-//   <React.StrictMode>
-//     <App />
-//   </React.StrictMode>,
-// )
+ ReactDOM.createRoot(document.getElementById('root')).render(
+   <React.StrictMode>
+     <App />
+   </React.StrictMode>,
+ )
 
-import React from 'react';
-import ReactDOM from 'react-dom/client';
-import ResultOverview from './ResultOverview.jsx'; // Import the new component
-import './index.css';
+//import React from 'react';
+//import ReactDOM from 'react-dom/client';
+//import ResultOverview from './ResultOverview.jsx'; // Import the new component
+//import './index.css';
 
-ReactDOM.createRoot(document.getElementById('root')).render(
-  <React.StrictMode>
-    <ResultOverview /> {/* Render the new component */}
-  </React.StrictMode>,
-);
+//ReactDOM.createRoot(document.getElementById('root')).render(
+//  <React.StrictMode>
+//    <ResultOverview /> {/* Render the new component */}
+//  </React.StrictMode>,
+//);
