+++ src/pages/Home.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+const Home = () => {
+    return (
+        <div>
+            <h1>Welcome to the Home Page</h1>
+            <p>This is the content of the home page.</p>
+        </div>
+    );
+};
+
+export default Home;
